adding Decimal numbers

+++ main.c
@@ -3,15 +3,15 @@
 
 int main() {
     
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     printf ("Enter first number: ");
-    scanf ("%d", &num1) ;
+    scanf ("%lf", &num1) ;
 
     printf ("Enter first number: ");
-    scanf ("%d", &num2);
+    scanf ("%lf", &num2);
 
-    printf ("Answer: %d", num1 + num2 );
+    printf ("Answer: %f", num1 + num2 );
 
     return 0;
 }
